requirements.txt
mpmath == 1.3.0
numpy == 1.24.3
pip == 23.0.1
pygame == 2.4.0
python-dotenv == 1.0.0